+++ VFS/source/ArcLA/Host.h
@@ -28,11 +28,11 @@ public:
     virtual bool IsImmutableFS() const noexcept override;
     
     virtual bool IsDirectory(const char *_path,
-                             int _flags,
+                             unsigned long _flags,
                              const VFSCancelChecker &_cancel_checker) override;
     
     virtual int StatFS(const char *_path, VFSStatFS &_stat, const VFSCancelChecker &_cancel_checker) override;
-    virtual int Stat(const char *_path, VFSStat &_st, int _flags, const VFSCancelChecker &_cancel_checker) override;
+    virtual int Stat(const char *_path, VFSStat &_st, unsigned long _flags, const VFSCancelChecker &_cancel_checker) override;
     
     virtual int CreateFile(const char* _path,
                            shared_ptr<VFSFile> &_target,
@@ -40,7 +40,7 @@ public:
     
     virtual int FetchDirectoryListing(const char *_path,
                                       shared_ptr<VFSListing> &_target,
-                                      int _flags,
+                                      unsigned long _flags,
                                       const VFSCancelChecker &_cancel_checker) override;
     
     virtual int IterateDirectoryListing(const char *_path, const function<bool(const VFSDirEnt &_dirent)> &_handler) override;
@@ -68,7 +68,7 @@ public:
     shared_ptr<ArchiveHost> SharedPtr() {return static_pointer_cast<ArchiveHost>(Host::SharedPtr());}
     
     /** return VFSError, not uids returned */
-    int ResolvePathIfNeeded(const char *_path, char *_resolved_path, int _flags);
+    int ResolvePathIfNeeded(const char *_path, char *_resolved_path, unsigned long _flags);
     
     enum class SymlinkState
     {
