+++ VFS/source/Native/Host.h
@@ -18,16 +18,16 @@ public:
     
     virtual int StatFS(const char *_path, VFSStatFS &_stat, const VFSCancelChecker &_cancel_checker) override;
     
-    virtual int Stat(const char *_path, VFSStat &_st, int _flags, const VFSCancelChecker &_cancel_checker) override;
+    virtual int Stat(const char *_path, VFSStat &_st, unsigned long _flags, const VFSCancelChecker &_cancel_checker) override;
     
     virtual int FetchDirectoryListing(const char *_path,
                                       shared_ptr<VFSListing> &_target,
-                                      int _flags,
+                                      unsigned long _flags,
                                       const VFSCancelChecker &_cancel_checker) override;
     
     virtual int FetchSingleItemListing(const char *_path_to_item,
                                        shared_ptr<VFSListing> &_target,
-                                       int _flags,
+                                       unsigned long _flags,
                                        const VFSCancelChecker &_cancel_checker) override;
     
     virtual int IterateDirectoryListing(const char *_path, const function<bool(const VFSDirEnt &_dirent)> &_handler) override;
