+++ VFS/source/NetWebDAV/WebDAVHost.h
@@ -33,14 +33,14 @@ public:
     
     int FetchDirectoryListing(const char *_path,
                               shared_ptr<VFSListing> &_target,
-                              int _flags,
+                              unsigned long _flags,
                               const VFSCancelChecker &_cancel_checker) override;
     
     int IterateDirectoryListing(const char *_path,
                                 const function<bool(const VFSDirEnt &_dirent)> &_handler) override;
     int Stat(const char *_path,
              VFSStat &_st,
-             int _flags,
+             unsigned long _flags,
              const VFSCancelChecker &_cancel_checker) override;
     
     int StatFS(const char *_path,
