+++ VFS/source/PS/Host.h
@@ -22,10 +22,10 @@ public:
                            const VFSCancelChecker &_cancel_checker) override;
     
     virtual bool IsDirectory(const char *_path,
-                             int _flags,
+                             unsigned long _flags,
                              const VFSCancelChecker &_cancel_checker) override;
     
-    virtual int Stat(const char *_path, VFSStat &_st, int _flags, const VFSCancelChecker &_cancel_checker) override;
+    virtual int Stat(const char *_path, VFSStat &_st, unsigned long _flags, const VFSCancelChecker &_cancel_checker) override;
     
     virtual int StatFS(const char *_path, VFSStatFS &_stat, const VFSCancelChecker &_cancel_checker) override;
     
@@ -33,7 +33,7 @@ public:
     
     virtual int FetchDirectoryListing(const char *_path,
                                       shared_ptr<VFSListing> &_target,
-                                      int _flags,
+                                      unsigned long _flags,
                                       const VFSCancelChecker &_cancel_checker) override;
     
     virtual int IterateDirectoryListing(const char *_path, const function<bool(const VFSDirEnt &_dirent)> &_handler) override;
